# Optimize binary packed encoding