--- a/encoding/delta/binary_packed.go
+++ b/encoding/delta/binary_packed.go
@@ -35,27 +35,6 @@ var (
 	encodeInt64 = encodeInt64Default
 )
 
-func miniBlockPackInt64(dst []byte, src *[miniBlockSize]int64, bitWidth uint) {
-	bitMask := uint64(1<<bitWidth) - 1
-	bitOffset := uint(0)
-
-	for _, value := range src {
-		i := bitOffset / 64
-		j := bitOffset % 64
-
-		lo := binary.LittleEndian.Uint64(dst[(i+0)*8:])
-		hi := binary.LittleEndian.Uint64(dst[(i+1)*8:])
-
-		lo |= (uint64(value) & bitMask) << j
-		hi |= (uint64(value) >> (64 - j))
-
-		binary.LittleEndian.PutUint64(dst[(i+0)*8:], lo)
-		binary.LittleEndian.PutUint64(dst[(i+1)*8:], hi)
-
-		bitOffset += bitWidth
-	}
-}
-
 func encodeInt32Default(dst []byte, src []int32) []byte {
 	totalValues := len(src)
 	firstValue := int32(0)