… encoding/delta/binary_packed_amd64.s
… encoding/delta/binary_packed_amd64_test.go
… encoding/delta/binary_packed_purego.go
… encoding/delta/binary_packed_purego_test.go
… encoding/delta/byte_array.go
… encoding/delta/byte_array_test.go
… encoding/delta/length_byte_array.go
… encoding/plain/plain.go
… encoding/plain/plain_amd64.go
… encoding/plain/plain_amd64.s
… encoding/plain/plain_purego.go
… encoding/plain/plain_test.go
… encoding/test/test.go
… encoding/test/test_go17.go
… encoding/test/test_go18.go
… parquet_test.go