This PR optimizes the encoding of values with DELTA_BINARY_PACKED.

These optimizations should have a wide effect since the encoding is used for DELTA_BYTE_ARRAY and DELTA_LENGTH_BYTE_ARRAY, the latter being the default encoding for string value in parquet.

The optimization comes in two forms:
* improve the pure Go implementation by working on 32 or 64 bits values instead of setting individual bits, making the algorithm O(numValues) instead of O(numValues*bitWidth)
* when available, leverage SIMD instructions of AVX2 to increase throughput further

# Description

The parquet format claims that delta encoding of integer values is derived from [Decoding billions of integers per second through vectorization](https://arxiv.org/pdf/1209.2137v5.pdf), tho after careful reading of the paper and related implementation in https://github.com/lemire/FastPFor, it appeared that there were many differences between the two.

First, encoding of values was not investigated in depth. The paper glanced over it and the FastPFor code had a trivial implementation. The paper focused on using SEE instructions, which left a lot of opportunities on the table to leverage AVX2 (e.g. using variable bit shifts of vector registers via VPSLLVD) and simplifies the algorithms (no need to write a custom version of for each bit width that values are getting packed into).

I ended having to reinvent an algorithm for generic vectorization of bit-packing, tho it applies only on bit width 0 to 16 (see `miniBlockPackInt32x3to16bits`). Since delta encoding is intended to encode values that are nearby in distance, I expect the current code to be effective in practice. Future research can be done to optimize delta encoding of larger deltas.

The parquet format differed in few other ways from the paper. For example, the paper uses vectorized prefix sum computation by performing deltas on groups of 4 integers, but the parquet format requires deltas to be computed between consecutive elements. As determined in the paper, after vectorizing the encoding of values, the next bottleneck was the computation of deltas, so I had to figure out an algorithm to optimize this process as well (see `blockDeltaInt32AVX2`).

The rest of the subroutines are optimizations that rely on the knowledge that blocks are of constant size. If we were to modify the block size in the future (or make it configurable), we could keep it a multiple of the current block size to be able to reuse the functions introduced in this PR.

Since there is a lot of added complexity here, I invested time in adding thorough testing, including to validate the behavior of subroutines of the algorithm.

# Benchmarks

I chose to invest time in optimizing DELTA_BINARY_PACKED because it showed up as the busiest code path and main bottleneck in many benchmarks, especially after the landing the optimizations from #199 

The following snapshots show the comparison of encoding throughput from the `main` branch:

> with `-tags=purego`
```
name                                             old time/op   new time/op    delta
Encode/DELTA_BINARY_PACKED/int32                   219µs ± 0%      69µs ± 0%   -68.42%  (p=0.000 n=10+9)
Encode/DELTA_BINARY_PACKED/int64                   229µs ±11%      70µs ± 0%   -69.38%  (p=0.000 n=10+9)
Encode/DELTA_LENGTH_BYTE_ARRAY/byte_array          352µs ± 0%     238µs ± 0%   -32.43%  (p=0.000 n=9+9)
Encode/DELTA_BYTE_ARRAY/byte_array                 690µs ± 0%     551µs ± 0%   -20.18%  (p=0.000 n=8+8)
Encode/DELTA_BYTE_ARRAY/fixed_length_byte_array    397µs ± 0%     359µs ± 0%    -9.58%  (p=0.000 n=10+9)

name                                             old speed     new speed      delta
Encode/DELTA_BINARY_PACKED/int32                 183MB/s ± 0%   578MB/s ± 0%  +216.68%  (p=0.000 n=10+9)
Encode/DELTA_BINARY_PACKED/int64                 350MB/s ±10%  1139MB/s ± 0%  +225.42%  (p=0.000 n=10+9)
Encode/DELTA_LENGTH_BYTE_ARRAY/byte_array        410MB/s ± 0%   607MB/s ± 0%   +48.00%  (p=0.000 n=9+9)
Encode/DELTA_BYTE_ARRAY/byte_array               210MB/s ± 0%   262MB/s ± 0%   +25.28%  (p=0.000 n=8+8)
Encode/DELTA_BYTE_ARRAY/fixed_length_byte_array  403MB/s ± 0%   445MB/s ± 0%   +10.59%  (p=0.000 n=10+9)

name                                             old value/s   new value/s    delta
Encode/DELTA_BINARY_PACKED/int32                   45.6M ± 0%    144.5M ± 0%  +216.68%  (p=0.000 n=10+9)
Encode/DELTA_BINARY_PACKED/int64                   43.8M ±10%    142.4M ± 0%  +225.42%  (p=0.000 n=10+9)
Encode/DELTA_LENGTH_BYTE_ARRAY/byte_array          28.4M ± 0%     42.0M ± 0%   +48.00%  (p=0.000 n=9+9)
Encode/DELTA_BYTE_ARRAY/byte_array                 14.5M ± 0%     18.2M ± 0%   +25.28%  (p=0.000 n=8+8)
Encode/DELTA_BYTE_ARRAY/fixed_length_byte_array    25.2M ± 0%     27.8M ± 0%   +10.58%  (p=0.000 n=10+9)
```

> with AVX2 optimized code paths
```
name                                             old time/op   new time/op    delta
Encode/DELTA_BINARY_PACKED/int32                   219µs ± 0%      10µs ± 0%    -95.57%  (p=0.000 n=10+8)
Encode/DELTA_BINARY_PACKED/int64                   229µs ±11%      29µs ± 0%    -87.57%  (p=0.000 n=10+10)
Encode/DELTA_LENGTH_BYTE_ARRAY/byte_array          352µs ± 0%     183µs ± 0%    -48.16%  (p=0.000 n=9+10)
Encode/DELTA_BYTE_ARRAY/byte_array                 690µs ± 0%     445µs ± 0%    -35.55%  (p=0.000 n=8+9)
Encode/DELTA_BYTE_ARRAY/fixed_length_byte_array    397µs ± 0%     277µs ± 0%    -30.21%  (p=0.000 n=10+8)

name                                             old speed     new speed      delta
Encode/DELTA_BINARY_PACKED/int32                 183MB/s ± 0%  4118MB/s ± 0%  +2155.74%  (p=0.000 n=10+8)
Encode/DELTA_BINARY_PACKED/int64                 350MB/s ±10%  2807MB/s ± 0%   +701.72%  (p=0.000 n=10+10)
Encode/DELTA_LENGTH_BYTE_ARRAY/byte_array        410MB/s ± 0%   792MB/s ± 0%    +92.90%  (p=0.000 n=9+10)
Encode/DELTA_BYTE_ARRAY/byte_array               210MB/s ± 0%   325MB/s ± 0%    +55.17%  (p=0.000 n=8+9)
Encode/DELTA_BYTE_ARRAY/fixed_length_byte_array  403MB/s ± 0%   577MB/s ± 0%    +43.29%  (p=0.000 n=10+8)

name                                             old value/s   new value/s    delta
Encode/DELTA_BINARY_PACKED/int32                   45.6M ± 0%   1029.5M ± 0%  +2155.76%  (p=0.000 n=10+8)
Encode/DELTA_BINARY_PACKED/int64                   43.8M ±10%    350.9M ± 0%   +701.71%  (p=0.000 n=10+10)
Encode/DELTA_LENGTH_BYTE_ARRAY/byte_array          28.4M ± 0%     54.7M ± 0%    +92.90%  (p=0.000 n=9+10)
Encode/DELTA_BYTE_ARRAY/byte_array                 14.5M ± 0%     22.5M ± 0%    +55.17%  (p=0.000 n=8+9)
Encode/DELTA_BYTE_ARRAY/fixed_length_byte_array    25.2M ± 0%     36.1M ± 0%    +43.29%  (p=0.000 n=10+8)
```

```
name                                          old time/op  new time/op  delta
GenericWriter/benchmarkRowType/go1.18         7.46µs ± 2%  7.37µs ± 1%    -1.23%  (p=0.007 n=10+10)
GenericWriter/booleanColumn/go1.18            1.24µs ± 4%  1.20µs ± 0%    -2.89%  (p=0.000 n=10+8)
GenericWriter/int32Column/go1.18              23.5µs ± 2%   2.4µs ± 0%   -89.78%  (p=0.000 n=8+9)
GenericWriter/int64Column/go1.18              2.26µs ± 0%  2.25µs ± 0%    -0.71%  (p=0.000 n=8+10)
GenericWriter/floatColumn/go1.18              2.01µs ± 0%  1.99µs ± 1%    -0.93%  (p=0.000 n=10+10)
GenericWriter/doubleColumn/go1.18             2.55µs ± 0%  2.56µs ± 1%      ~     (p=0.670 n=10+10)
GenericWriter/byteArrayColumn/go1.18          60.3µs ± 0%  46.1µs ± 1%   -23.48%  (p=0.000 n=9+10)
GenericWriter/fixedLenByteArrayColumn/go1.18  15.3µs ± 1%  15.3µs ± 1%      ~     (p=0.782 n=10+10)
GenericWriter/stringColumn/go1.18             36.6µs ± 1%  32.0µs ± 0%   -12.48%  (p=0.000 n=9+9)
GenericWriter/indexedStringColumn/go1.18       114µs ± 1%   113µs ± 1%    -0.57%  (p=0.028 n=9+10)
GenericWriter/uuidColumn/go1.18               44.3µs ± 0%  33.6µs ± 1%   -24.22%  (p=0.000 n=10+10)
GenericWriter/mapColumn/go1.18                 894µs ± 2%   891µs ± 2%      ~     (p=0.393 n=10+10)
GenericWriter/decimalColumn/go1.18            2.26µs ± 1%  2.24µs ± 0%    -0.51%  (p=0.017 n=10+10)
GenericWriter/contact/go1.18                  49.7µs ± 1%  49.6µs ± 1%      ~     (p=0.529 n=10+10)

name                                          old row/s    new row/s    delta
GenericWriter/benchmarkRowType/go1.18           134M ± 2%    136M ± 1%    +1.24%  (p=0.007 n=10+10)
GenericWriter/booleanColumn/go1.18              807M ± 4%    830M ± 0%    +2.89%  (p=0.000 n=10+9)
GenericWriter/int32Column/go1.18               42.6M ± 2%  416.4M ± 0%  +878.06%  (p=0.000 n=8+9)
GenericWriter/int64Column/go1.18                442M ± 0%    445M ± 0%    +0.73%  (p=0.000 n=8+10)
GenericWriter/floatColumn/go1.18                499M ± 0%    503M ± 1%    +0.94%  (p=0.000 n=10+10)
GenericWriter/doubleColumn/go1.18               392M ± 0%    391M ± 1%      ~     (p=0.739 n=10+10)
GenericWriter/byteArrayColumn/go1.18           16.6M ± 0%   21.7M ± 1%   +30.69%  (p=0.000 n=9+10)
GenericWriter/fixedLenByteArrayColumn/go1.18   65.4M ± 1%   65.3M ± 1%      ~     (p=0.796 n=10+10)
GenericWriter/stringColumn/go1.18              27.3M ± 1%   31.2M ± 0%   +14.25%  (p=0.000 n=9+9)
GenericWriter/indexedStringColumn/go1.18       8.77M ± 1%   8.82M ± 1%    +0.57%  (p=0.028 n=9+10)
GenericWriter/uuidColumn/go1.18                22.6M ± 0%   29.8M ± 1%   +31.96%  (p=0.000 n=10+10)
GenericWriter/mapColumn/go1.18                 1.12M ± 2%   1.12M ± 2%      ~     (p=0.393 n=10+10)
GenericWriter/decimalColumn/go1.18              443M ± 1%    446M ± 0%    +0.51%  (p=0.023 n=10+10)
GenericWriter/contact/go1.18                   20.1M ± 1%   20.2M ± 1%      ~     (p=0.529 n=10+10)
```

Please take a look and let me know if you have any questions.