@achille-roussel 2022-06-01
The encodings are part of the parquet spec and documented at https://github.com/apache/parquet-format/blob/master/Encodings.md


This seems to be a similar to the protobuf documents. Do you think it is useful or did you picture something different?