@kevinburkesegment 2022-06-01
Ah, I assumed from "encoding of values was not investigated in depth" and "ended having to reinvent an algorithm for generic vectorization of bit-packing" that the packed format was left unspecified and you added new optimizations there. Maybe not.