@achille-roussel 2022-06-01
The parquet spec specifies how the bits need to be laid out, but we have some flexibility on how to write the program that generates it. I tried to document details about the implementation choices in the source code, the rest should be found in the public specification of parquet encodings.